+++ CHANGELOG.md
@@ -34,6 +34,8 @@ Bearbeitung mit RexStan und Bereinigung diverser Fehler. Das alles führt zu ein
 - Neu:
   - Test der Layer-URLs interaktiv im Eingabeformular (#100)
   - Individuelles CSS (`redaxo/data/geolocation/geolocation.css') kann auch in SCSS-Dateien stehen (Editor-frundlich) (#104)
+  - Für Basiskarten im Kartensatz/Mapset kann die aktive Karte unabhängig von der Reihenfolge (bisher immer die erste) ber RAdio-Button aktiviert werden (#107)
+  - Für Overlay-Karten im Kartensatz/Mapset können sofort sichtbare Overlays aktiviert werden (Checkbox); bisher waren die Karten immer initial ausgeblendet (#107)
 
 ## 06.05.2022 1.0.2